Added rows and columns to map size, added min for row and column in shcema
Added size to the map where rows and columns are entered in the xml, and
the schema now has a min value for those so the map has to have size at
least 8x12, changed each map from beeing String array to a
ArrayList<String> to support undefined size. meaning that the
ArrayList<String []> now is a ArrayList<ArrayList<String>>, fixed test
to support the new type. Tested the map with parserTest and it looks
like it should.