Created the reflection loading of zone classes
Created a zone class BearTrap for testing purposes but also works in a
game. Changed the xml schema so that there can be no zone class or one
for each level. fixed the levelbuilder to take the reflection loaded
class and method. Changed level to contain the reflection loaded class
and method, game changed to add the reflection loaded class and method
to each roadTile.